sync-users/requirements.txt
faker==24.4.0
fastapi==0.110.0
dispatch-py==0.4.0
requests==2.31.0
uvicorn==0.29.0
